requirements.txt
pandas
scikit-learn
matplotlib
sentence-transformers
fastapi
uvicorn[standard]
colorama
python-dotenv 